Fix inline function doesn't have coverage data.
If code under test are compiled as a library with coverage instrumentation,
and test code are compiled without coverage instrumentation, inline function
in header of the library WON'T generate coverage data when called from test code.
That's because the inline function in the library is invisible from test code,
and the one called by test code is another version of the inline function
defined in test code's translation unit instead of the instrumented function
in the library.